refactor component filenames to kebab-case and update calendar icon
