+++ app/(app)/app/page.tsx
@@ -1,5 +1,5 @@
 "use client"
-import Calendar from "@/components/Calendar";
+import Calendar from "@/components/calendar";
 
 export default function Home() {
   return <Calendar />;
