+++ app/(auth)/reset-password/page.tsx
@@ -1,6 +1,6 @@
 import { Calendar } from "lucide-react"
 
-import { ResetPasswordForm } from "@/components/account/ResetForm";
+import { ResetPasswordForm } from "@/components/account/reset-form";
 
 export default function ResetPasswordPage() {
   return (
