+++ app/(auth)/sign-in/page.tsx
@@ -1,5 +1,5 @@
 import { Calendar } from "lucide-react"
-import { LoginForm } from "@/components/account/LoginForm"
+import { LoginForm } from "@/components/account/login-form"
 
 export default function LoginPage() {
   return (
