+++ components/analytics/build-info-card.tsx
