+++ components/calendar.tsx
@@ -6,25 +6,25 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue, SelectGr
 import { Input } from "@/components/ui/input"
 import { ChevronLeft, ChevronRight, Search, PanelLeft, BarChart2, Settings as SettingsIcon } from 'lucide-react'
 import { addDays, addYears, subDays, subYears } from "date-fns"
-import Sidebar from "@/components/sidebar/Sidebar"
-import DayView from "@/components/view/DayView"
-import WeekView from "@/components/view/WeekView"
-import MonthView from "@/components/view/MonthView"
-import YearView from "@/components/view/YearView"
-import EventDialog from "@/components/event/EventDialog"
-import Settings from "@/components/home/Settings"
-import UserProfileButton, { type UserProfileSection } from "@/components/home/UserProfileButton"
+import Sidebar from "@/components/sidebar/sidebar"
+import DayView from "@/components/view/day-view"
+import WeekView from "@/components/view/week-view"
+import MonthView from "@/components/view/month-view"
+import YearView from "@/components/view/year-view"
+import EventDialog from "@/components/event/event-dialog"
+import Settings from "@/components/home/settings"
+import UserProfileButton, { type UserProfileSection } from "@/components/home/user-profile-button"
 import { translations, useLanguage } from "@/lib/i18n"
 import { checkPendingNotifications, clearAllNotificationTimers, type NOTIFICATION_SOUNDS } from "@/lib/notifications"
-import EventPreview from "@/components/event/EventPreview"
+import EventPreview from "@/components/event/event-preview"
 import { readEncryptedLocalStorage, useLocalStorage, writeEncryptedLocalStorage } from "@/hooks/useLocalStorage"
-import { useCalendar } from "@/components/context/CalendarContext"
-import EventUrlHandler from "@/components/event/EventUrlHandler"
-import RightSidebar from "@/components/sidebar/RightSidebar"
-import AnalyticsView from "@/components/analytics/AnalyticsView"
+import { useCalendar } from "@/components/context/calendar-context"
+import EventUrlHandler from "@/components/event/event-url-handler"
+import RightSidebar from "@/components/sidebar/right-sidebar"
+import AnalyticsView from "@/components/analytics/analytics-view"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { cn } from "@/lib/utils"
-import DailyToast from "@/components/home/DailyToast"
+import DailyToast from "@/components/home/daily-toast"
 import { toast } from "sonner"
 import {
   AlertDialog,
