+++ components/event/event-url-handler.tsx
@@ -13,7 +13,7 @@ import {
   AlertDialogTitle,
 } from "@/components/ui/alert-dialog"
 import { addDays } from "date-fns"
-import { useCalendar } from "@/components/context/CalendarContext"
+import { useCalendar } from "@/components/context/calendar-context"
 import { isZhLanguage, translations, useLanguage } from "@/lib/i18n"
 import { toast } from "sonner"
