+++ components/home/shared-event.tsx
@@ -25,7 +25,7 @@ import { cn } from "@/lib/utils";
 import { isZhLanguage, useLanguage } from "@/lib/i18n";
 
 import { Button } from "@/components/ui/button";
-import { useCalendar } from "@/components/context/CalendarContext";
+import { useCalendar } from "@/components/context/calendar-context";
 import { motion } from "framer-motion";
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent, CardDescription, CardTitle } from "@/components/ui/card";
