+++ components/view/month-view.tsx
@@ -3,7 +3,7 @@
 import { format, startOfMonth, endOfMonth, eachDayOfInterval, isSameMonth, isSameDay, subDays } from "date-fns"
 import { cn } from "@/lib/utils"
 import { translations, type Language } from "@/lib/i18n"
-import type { CalendarEvent } from "../Calendar"
+import type { CalendarEvent } from "../calendar"
 
 interface MonthViewProps {
   date: Date
